requirements.txt
annotated-types==0.5.0
anyio==3.7.1
blinker==1.7.0
certifi==2023.7.22
click==8.1.7
distro==1.8.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
openai==1.2.4
packaging==23.2
pydantic==2.5.0
pydantic_core==2.14.1
sniffio==1.3.0
SQLAlchemy==2.0.23
tqdm==4.66.1
typing_extensions==4.8.0
Werkzeug==3.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9